subject.cc

+++ Jim-Scott-Neo/block.cc
@@ -80,8 +80,8 @@ bool Block::isSuccessful() {
 // protected
 void Block::modifyCellsUnderGrid(bool is_erase)const {
 
-    for (int r = 0;r < small_grid.capacity(); r++) {
-        for (int c = 0;c < small_grid.at(0).capacity(); c++) {
+    for (int r = 0;r < (int) small_grid.capacity(); r++) {
+        for (int c = 0;c < (int) small_grid.at(0).capacity(); c++) {
             if (small_grid.at(r).at(c) == my_type) {
                 if (is_erase)cellAt(anchor, c, r)->setPiece(WhoIam::Null, isNull);
                 else cellAt(anchor, c, r)->setPiece(my_type, level);
